Engine audit: fix lint debt, split hysteresis tests, add equilibrium invariant
Findings from a full audit of the engine calculations and tests. The suite,
typecheck, build and all 33 validation benchmarks were already clean; these are
the three real gaps it turned up, all of them mine from this session.

1. hysteresis.ts had 5 exported functions but NO matching *.test.ts — its cases
   were living inside nonlinearTimeHistory.test.ts, breaking the repo rule that
   every engine module carries its own test file. Split them into
   hysteresis.test.ts (unchanged assertions). The only engine modules now
   without a test file are type-only (model, schedule/model, progress) or
   Worker/pool infrastructure that cannot be unit-tested (framePool,
   frameWorker, solverWorker).

2. Five lint errors I introduced: prefer-const on u/v in nonlinearTimeHistory
   (mutated in place, never reassigned) and on minX/maxX/slicePts in
   SlopeStability.tsx. Repo lint total 44 -> 39 errors; the remainder is
   pre-existing debt the CI workflow already documents as non-blocking.

3. axialOnly had no equilibrium check. Deactivating a member must not leave
   unbalanced force, and that is the single most important invariant of an
   active-set solve, so it is now asserted permanently rather than spot-checked:
   sum of reactions = applied load to 1e-9 in both load directions and at two
   magnitudes, plus a linearity guard (within one unchanged active set the
   response must scale exactly, proving the iteration injects no spurious state
   between solves).

Audit results that needed no action: 1533 tests passing, tsc -b and build clean,
33 validation cases all finite and under 1% deviation with the worst consuming
only 4% of its tolerance budget, no skipped/only/todo tests anywhere, no `any`
in engine code, and the L3 statics self-checks (reaction equilibrium, closed-form
cantilever, fixed-end moments) present in frame3d, frame2d, modelBridge, truss
and beamAnalysis.

Suite 1533 -> 1535.
